candidate management file splited
