+++ Frontend/apps/interview/src/hooks/index.ts
@@ -2,3 +2,4 @@
 // Export hooks from this file as they are created
 export { useWizard } from "./use-wizard";
 export { useTestFilters } from "./use-test-filters";
+export { useNetworkStatus } from "./use-network-status";
